DB multithreading/readme/fixed bugs

+++ server/sqlite.py
@@ -1,83 +1,89 @@
 import sqlite3
 
+from threading import Lock
+
 
 class DatabaseException(Exception):
     pass
+
+
+def create_select_statement(table_name: str, fields: list[str] = None) -> str:
+    """
+    Create SELECT-statement from field list
+
+    :param table_name: Name of the table
+    :param fields: List of the fields names
+
+    :return: SELECT-clause
+    """
+    query = 'SELECT '
+    if fields is None:
+        query += "*"
+    else:
+        for idx, field in enumerate(fields):
+            if idx == 0:
+                query += f'"{field}"'
+            else:
+                query += f', "{field}"'
+
+    query += f' FROM {table_name} '
+
+    return query
+
+
+def create_where_statement(where: dict[str, str]) -> str:
+    """
+    Create WHERE-statement from dictionary
+
+    :param where: Dictionary with WHERE-parameters
+
+    :return: WHERE-clause
+    """
+    query = 'WHERE '
+    for idx, param in enumerate(where.keys()):
+        value = where[param]
+
+        if idx == 0:
+            query += f'"{param}" = "{value}"'
+        else:
+            query += f' AND "{param}" = "{value}"'
+
+    return query
 
 
 class SQLite:
     """
     Class to working with SQLite databases
     """
-    def __init__(self, dbfile):
+    def __init__(self):
         """
         Initialize SQLite
 
         :param dbfile: Full path to the .db file
         """
-        self.__dbfile = dbfile
+        self.__conn = None
+        self.__cursor = None
+        self.__lock = Lock()
 
-    def __enter__(self):
+    def open(self, file_name: str) -> None:
         """
-        Open database connection
+        Open database
         """
         try:
-            self.__conn = sqlite3.connect(self.__dbfile)
+            self.__conn = sqlite3.connect(file_name)
             self.__cursor = self.__conn.cursor()
         except Exception as e:
             raise DatabaseException(f'SQLite connection error: {e}')
 
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def close(self) -> None:
         """
         Close database connection
         """
-        self.__conn.close()
+        with self.__lock:
+            self.__conn.close()
 
-    def __create_select_statement(self, table_name: str, fields: list[str] = []) -> str:
-        """
-        Create SELECT-statement from field list
-
-        :param table_name: Name of the table
-        :param fields: List of the fields
-
-        :return: SELECT-clause
-        """
-        query = 'SELECT '
-        if len(fields) == 0:
-            query += "*"
-        else:
-            for idx, field in enumerate(fields):
-                if idx == 0:
-                    query += f'"{field}"'
-                else:
-                    query += f', "{field}"'
-
-        query += f' FROM {table_name} '
-
-        return query
-
-    def __create_where_statement(self, where: dict[str, str]) -> str:
-        """
-        Create WHERE-statement from dictionary
-
-        :param where: Dictionary with WHERE-parameters
-
-        :return: WHERE-clause
-        """
-        query = 'WHERE '
-        for idx, param in enumerate(where.keys()):
-            value = where[param]
-
-            if idx == 0:
-                query += f'"{param}" = "{value}"'
-            else:
-                query += f' AND "{param}" = "{value}"'
-
-        return query
-
-    def select(self, table_name: str, fields: list[str] = None, where: dict[str, str] = None, orderby: str = None, groupby: str = None) -> list:
+    def select(self, table_name: str, fields: list[str] = None, where: dict[str, str] = None,
+               orderby: str = None, groupby: str = None) -> list:
         """
         Select records from table
 
@@ -88,15 +94,21 @@ class SQLite:
         :param groupby: Field name to group records
         :return:
         """
-        try:
-            if fields is None:
-                self.__cursor.execute(f'PRAGMA table_info("{table_name}")')
-                fields = [entry[1] for entry in self.__cursor.fetchall()]
+        if self.__conn is None or self.__cursor is None:
+            raise DatabaseException('No opened database')
 
-            query = self.__create_select_statement(table_name, fields)
+        try:
+            result = []
+
+            if fields is None:
+                with self.__lock:
+                    self.__cursor.execute(f'PRAGMA table_info("{table_name}")')
+                    fields = [entry[1] for entry in self.__cursor.fetchall()]
+
+            query = create_select_statement(table_name, fields)
 
             if where is not None:
-                query += self.__create_where_statement(where)
+                query += create_where_statement(where)
 
             if orderby is not None:
                 query += f' ORDER BY {orderby}'
@@ -104,21 +116,19 @@ class SQLite:
             if groupby is not None:
                 query += f' GROUP BY {groupby}'
 
-            self.__cursor.execute(query)
+            with self.__lock:
+                self.__cursor.execute(query)
+                db_result = self.__cursor.fetchall()
 
-            db_result = self.__cursor.fetchall()
-
-            result = []
             for db_entry in db_result:
                 entry = {}
                 for idx, field in enumerate(fields):
                     entry[field] = db_entry[idx]
+
                 result.append(entry)
 
             return result
         except Exception as e:
-            print(e)
-            print(type(e))
             raise DatabaseException(f'SQLite error: {e}')
 
     def insert(self, table_name: str, values: dict[str, str]) -> int:
@@ -128,6 +138,9 @@ class SQLite:
         :param table_name: Name of the table
         :param values: Dictionary with values to insert
         """
+        if self.__conn is None or self.__cursor is None:
+            raise DatabaseException('No opened database')
+
         try:
             query = f'INSERT INTO {table_name}('
             insert_values = ''
@@ -142,21 +155,25 @@ class SQLite:
 
             query += f') VALUES ({insert_values})'
 
-            self.__cursor.execute(query)
-            self.__conn.commit()
+            with self.__lock:
+                self.__cursor.execute(query)
+                self.__conn.commit()
 
-            return self.__cursor.lastrowid
+                return self.__cursor.lastrowid
         except Exception as e:
             raise DatabaseException(f'SQLite error: {e}')
 
     def update(self, table_name: str, values: dict[str, str], where: dict[str, str] = None) -> None:
         """
-        Update table from a dictonary values
+        Update table from a dictionary values
 
         :param table_name: Name of the table
         :param values: Dictionary with values to update
         :param where: Dictionary with WHERE parameters
         """
+        if self.__conn is None or self.__cursor is None:
+            raise DatabaseException('No opened database')
+
         try:
             query = f'UPDATE "{table_name}" SET '
 
@@ -168,9 +185,10 @@ class SQLite:
                     query += f', "{param}" = "{value}"'
 
             if where is not None:
-                query += ' ' + self.__create_where_statement(where)
+                query += ' ' + create_where_statement(where)
 
-            self.__cursor.executescript(query)
+            with self.__lock:
+                self.__cursor.executescript(query)
         except Exception as e:
             raise DatabaseException(f'SQLite error: {e}')
 
@@ -181,36 +199,34 @@ class SQLite:
         :param table_name: Name of the table
         :param where: Dictionary with WHERE parameters
         """
+        if self.__conn is None or self.__cursor is None:
+            raise DatabaseException('No opened database')
+
         try:
             query = f'DELETE FROM "{table_name}" ';
 
             if where is not None:
-                query += self.__create_where_statement(where)
+                query += create_where_statement(where)
 
-            self.__cursor.executescript(query)
+            with self.__lock:
+                self.__cursor.executescript(query)
         except Exception as e:
             raise DatabaseException(f'SQLite error: {e}')
 
     def custom_select(self, query: str, commit: bool = False) -> None:
         """
-        Perforum custom SELECT-request
+        Perform custom SELECT-request
 
         :param query: SQL query
         :param commit: True to run SQL COMMIT-command
         """
+        if self.__conn is None or self.__cursor is None:
+            raise DatabaseException('No opened database')
+
         try:
-            self.__logger.debug(f'Execute query: {query}')
-            self.__cursor.executescript(query)
-            if commit:
-                self.__conn.commit()
+            with self.__lock:
+                self.__cursor.executescript(query)
+                if commit:
+                    self.__conn.commit()
         except Exception as e:
             raise DatabaseException(f'SQLite error: {e}')
-
-    def start_transaction(self):
-        self.__cursor.execute('BEGIN TRANSACTION')
-
-    def commit_transaction(self):
-        self.__cursor.execute('COMMIT')
-
-    def rollback_transaction(self):
-        self.__cursor.execute('ROLLBACK')
